Apply restructure to signature widget
